done adding logic to switch between prompting modes, still have to implement verification
