+++ package.json
@@ -8,6 +8,7 @@
     "dotenv": "^16.4.1",
     "fastest-levenshtein": "^1.0.16",
     "flat": "5.0.2",
+    "gemini-zod": "^0.1.2",
     "iso-639-1": "^3.1.0",
     "ollama": "^0.5.12",
     "openai": "^4.27.1",
