+++ src/constants.ts
@@ -23,6 +23,8 @@ export const CLI_HELP = {
     OllamaHost:
         "The host and port number serving Ollama. 11434 is the default port number.",
     OverridePromptFile: `Use the prompts from the given JSON file, containing keys for at least one of ${OVERRIDE_PROMPT_KEYS.join(", ")}`,
+    PromptMode:
+        "Chose the prompting mode, between 'csv' mode (better performance but will only work with advanced models like GPT-40) or 'json' mode (compatible with less avanced models, like llama:8b, but translations take 50% longer on average)",
     RateLimit:
         "How many milliseconds between requests (defaults to 1s for Gemini, 120ms (at 500RPM) for ChatGPT, 1200ms for Claude)",
     SkipStylingVerification:
