+++ src/enums/prompt_mode.ts
@@ -1,0 +1,6 @@
+enum PromptMode {
+    CSV = "csv",
+    JSON = "json",
+}
+
+export default PromptMode;
