+++ src/generate_json/generate_json.ts
@@ -1,13 +1,32 @@
+import * as cl100k_base from "tiktoken/encoders/cl100k_base.json";
+import { DEFAULT_BATCH_SIZE, MAX_TOKEN } from "src/constants";
+import { Tiktoken } from "tiktoken";
 import { TranslateItemOutputObjectSchema } from "./types_json";
 import { generationPromptJson } from "./prompts_json";
 import { retryJob } from "src/utils";
-import type { GenerateState } from "src/types";
+import type { GenerateState, TranslationStats } from "src/types";
 import type {
     TranslateItem,
     TranslateItemInput,
     TranslateItemOutput,
 } from "./types_json";
+import type Chats from "src/interfaces/chats";
 import type GenerateTranslationOptionsJson from "src/interfaces/generate_translation_options_json";
+import type TranslateOptions from "src/interfaces/translate_options";
+
+function createJsonInput(
+    id: number,
+    key: string,
+    original: string,
+): TranslateItem {
+    return {
+        id,
+        // context: "",
+        key,
+        original,
+        translated: "",
+    } as TranslateItem;
+}
 
 function getTranslateItemsInput(
     translateItems: TranslateItem[],
@@ -22,11 +41,155 @@ function getTranslateItemsInput(
     );
 }
 
+function getTokenCount(text: string): number {
+    const encoding = new Tiktoken(
+        cl100k_base.bpe_ranks,
+        cl100k_base.special_tokens,
+        cl100k_base.pat_str,
+    );
+
+    return encoding.encode(text).length;
+}
+
+function getBatchTranslateItemArray(
+    translateItemArray: TranslateItem[],
+    options: TranslateOptions,
+): TranslateItem[] {
+    const promptTokens = getTokenCount(
+        generationPromptJson(
+            options.inputLanguage,
+            options.outputLanguage,
+            [],
+            options.overridePrompt,
+        ),
+    );
+
+    const maxInputTokens = ((MAX_TOKEN - promptTokens) * 0.9) / 2;
+
+    let currentTokens = 0;
+
+    const batchTranslateItemArray: TranslateItem[] = [];
+
+    for (const translateItem of translateItemArray) {
+        currentTokens += getTokenCount(
+            JSON.stringify(getTranslateItemsInput([translateItem])[0]),
+        );
+
+        if (
+            currentTokens >= maxInputTokens ||
+            batchTranslateItemArray.length >=
+                Number(options.batchSize ?? DEFAULT_BATCH_SIZE)
+        ) {
+            break;
+        }
+
+        batchTranslateItemArray.push(translateItem);
+    }
+
+    return batchTranslateItemArray;
+}
+
 /**
  * Complete the initial translation of the input text.
+ * @param flatInput - The flatinput object containing the json to translate
  * @param options - The options to generate the translation
+ * @param chats - The options to generate the translation
+ * @param translationStats - The translation statictics
  */
-export default async function generateTranslationJson(
+export default async function translateJson(
+    flatInput: { [key: string]: string },
+    options: TranslateOptions,
+    chats: Chats,
+    translationStats: TranslationStats,
+): Promise<{ [key: string]: string }> {
+    const translateItemArray: TranslateItem[] = [];
+
+    for (let i = 0; i < Object.keys(flatInput).length; i++) {
+        const key = Object.keys(flatInput)[i];
+        if (Object.prototype.hasOwnProperty.call(flatInput, key)) {
+            translateItemArray.push(
+                createJsonInput(i + 1, key, flatInput[key]),
+            );
+        }
+    }
+
+    const generatedTranslation: TranslateItem[] = [];
+    const totalItems = translateItemArray.length;
+
+    while (translateItemArray.length > 0) {
+        if (translationStats.processedItems > 0 && options.verbose) {
+            console.log(
+                `Completed ${((translationStats.processedItems / totalItems) * 100).toFixed(0)}%`,
+            );
+
+            const roundedEstimatedTimeLeftSeconds = Math.round(
+                (((Date.now() - translationStats.batchStartTime) /
+                    (translationStats.processedItems + 1)) *
+                    (totalItems - translationStats.processedItems)) /
+                    1000,
+            );
+
+            console.log(
+                `Estimated time left: ${roundedEstimatedTimeLeftSeconds} seconds`,
+            );
+        }
+
+        const batchTranslateItemArray = getBatchTranslateItemArray(
+            translateItemArray,
+            options,
+        );
+
+        translationStats.enqueuedItems += batchTranslateItemArray.length;
+
+        // eslint-disable-next-line no-await-in-loop
+        const result = await generateTranslationJson({
+            chats,
+            ensureChangedTranslation: options.ensureChangedTranslation ?? false,
+            inputLanguage: `[${options.inputLanguage}]`,
+            outputLanguage: `[${options.outputLanguage}]`,
+            overridePrompt: options.overridePrompt,
+            skipStylingVerification: options.skipStylingVerification ?? false,
+            skipTranslationVerification:
+                options.skipTranslationVerification ?? false,
+            templatedStringPrefix: options.templatedStringPrefix,
+            templatedStringSuffix: options.templatedStringSuffix,
+            translateItems: batchTranslateItemArray,
+            verboseLogging: options.verbose ?? false,
+        });
+
+        // console.log(result);
+        if (!result) {
+            console.error(
+                `Failed to generate translation for ${options.outputLanguage}`,
+            );
+            break;
+        }
+
+        for (const translatedItem of result) {
+            const index = translateItemArray.findIndex(
+                (item) => item.id === translatedItem.id,
+            );
+
+            if (index !== -1) {
+                translateItemArray.splice(index, 1);
+                generatedTranslation.push(translatedItem);
+            }
+
+            translationStats.processedItems++;
+        }
+    }
+
+    const output: { [key: string]: string } = {};
+
+    // Convert array of TranslateItem objects to output
+    for (const translation of generatedTranslation) {
+        output[translation.key] = translation.translated;
+    }
+
+    return output;
+}
+
+async function generateTranslationJson(
     options: GenerateTranslationOptionsJson,
 ): Promise<TranslateItem[]> {
     const generationPromptText = generationPromptJson(
