+++ src/interfaces/options.ts
@@ -1,6 +1,7 @@
 import type { ChatParams, Model } from "../types";
 import type Engine from "../enums/engine";
 import type OverridePrompt from "./override_prompt";
+import type PromptMode from "src/enums/prompt_mode";
 
 export default interface Options {
     engine: Engine;
@@ -17,4 +18,5 @@ export default interface Options {
     skipTranslationVerification?: boolean;
     skipStylingVerification?: boolean;
     overridePrompt?: OverridePrompt;
+    promptMode: PromptMode;
 }
