+++ src/types.ts
@@ -2,6 +2,7 @@ import type { ChatRequest } from "ollama";
 import type { ChatSession, StartChatParams } from "@google/generative-ai";
 import type { MessageCreateParams } from "@anthropic-ai/sdk/resources";
 import type OpenAI from "openai";
+import type PromptMode from "./enums/prompt_mode";
 
 export type Chat = ChatSession | OpenAI.Completion;
 
@@ -22,6 +23,7 @@ export type ModelArgs = {
     rateLimitMs: number;
     apiKey: string | undefined;
     host: string | undefined;
+    promptMode: PromptMode;
 };
 
 export type GenerateState = {
